Adapt OldT9Engine to new interface
- Fix trie build test giving false positive
